increased m parameter size in ex 4

+++ class_2/include/golomb_codec.h
@@ -65,7 +65,7 @@ class Predictor {
 
 struct Block {
     /* Header */
-    uint8_t m;  // increase this size
+    uint16_t m;  // increase this size
     PREDICTOR_TYPE predictor;
     /* Data */
     vector<short> data;
